Refactor inspector a bit
