+++ composer.json
@@ -61,7 +61,7 @@
             "source-directory": "config"
         },
         "config-plugin": {
-            "bootstrap": "bootstrap.php",
+            "bootstrap-web": "bootstrap-web.php",
             "params": "params.php",
             "web": "web.php",
             "routes": "routes.php",
