+++ config/bootstrap-web.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
  * @var $params array
  */
 
-use Yiisoft\Yii\Debug\Api\ApplicationState;
+use Yiisoft\Yii\Debug\Api\Inspector\ApplicationState;
 
 return [
     static function ($container) use ($params) {
